clippy lints: explicitly set truncate options when creating files
The disk cache should not truncate files on creation, since it is
responsible for writing out completed pieces to files; truncating would
destroy any already-completed pieces.

All the other places that request creation of the file should truncate
though; after opening/creating the file, they write out the entire
contents of the file. `create_new(true)` would be another option;
however, this might lead to synapse failing in surprising ways if it
previously died in the middle of trying to write out one of these files.