Only serialize session state if the state is dirty
synapse serializes session state periodically; that session state
includes per-torrent state. However, per-torrent state was serialized,
even if it hadn't changed, flooding the disk worker thread with needless
tasks.

Also add a fallback check to force serialization if the `dirty` bit is
incorrectly maintained. While somewhat inefficient, it should not matter
in practice, since double-serialization is only triggered if the `dirty`
bit is incorrect. In theory, this additional safety check can be removed
in the future with better test coverage.