Purge entries from the file cache after a torrent is moved
While these entries might be eventually pruned, leaving them around
means the live file descriptors will prevent the filesystem from
reclaiming the space. The cache cleanup is implemented as a separate
message to preserve ordering, as there may be in-flight requests to the
disk thread that reference files at the previous location.