Replace FFI `fallocate()` implementation with rustix
Also fix the `is_sparse()` implementation: on POSIX systems, `stat()`
returns the number of 512-byte blocks in `stat()`, so it does not make
sense to multiply it with the blocksize returned by `stat()`. Instead,
seek to the first hole to if a file is sparse.

It is unclear what notion of sparseness this function should check for;
the disk cache itself considers a file non-sparse if `fallocate()`
succeeds–but `is_sparse()` considers a file allocated with `fallocate()`
to be sparse until all the holes are filled in. It is possible this
check is not needed anymore: the original sparseness check was in commit
00c6b312470b0ce3c8c6a1aba948e4d1cd935885, and the `mmap()` logic has
since been removed.