Remove CONFIG global variable
Pass it around as an Arc<Config> instead. This removes the hack that was
previously needed to make sure the otherwise-lazy-loaded config was
actually loaded successfully.

There are a few places that don't really need a full Arc<Config> and
could be further simplified, e.g.:
- rpc::Processor only needs a path to the session directory
- rpc::Client::Incoming really only needs a &RpcConfig
But some of these get tricky to implement in a way that satisfies the
borrow checker without requiring unsafety, so just do the simple thing
for now.